pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tusk-drift-python-sdk"
version = "0.1.13"
description = "Python SDK for Tusk Drift instrumentation and replay"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Tusk", email = "support@usetusk.ai"}
]
keywords = ["tusk", "drift", "testing", "instrumentation", "tracing", "replay"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "protobuf>=6.0",
    "PyYAML>=6.0",
    "requests>=2.32.5",
    "tusk-drift-schemas>=0.1.24",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "time-machine>=2.10.0",
    "typing_extensions>=4.4.0",
]

[project.optional-dependencies]
flask = ["Flask>=3.1.2"]
fastapi = ["fastapi>=0.115.6", "uvicorn>=0.34.2", "starlette<0.42.0"]
django = ["Django>=4.2"]
dev = [
    "Flask>=3.1.2",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.2",
    "python-jsonpath>=0.10",
    "ruff>=0.8.0",
    "ty>=0.0.1a7",
    "pytest>=8.0.0,<9.0.0",
    "pytest-mock>=3.15.0",
    "pytest-cov>=7.0.0",
]

[project.urls]
Homepage = "https://usetusk.ai"
Documentation = "https://docs.usetusk.ai"
Repository = "https://github.com/Use-Tusk/drift-python-sdk"
Issues = "https://github.com/Use-Tusk/drift-python-sdk/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["drift*"]

[tool.setuptools.package-data]
drift = ["py.typed"]

[tool.uv]
prerelease = "allow"

[tool.ruff]
target-version = "py39"
line-length = 120
exclude = [
    # Ignore all rules for generated protobuf and grpc
    "**/greeter_pb2.py",
    "**/greeter_pb2_grpc.py",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module import not at top of file (common for conditional imports)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "ARG001", # unused function argument (common in handlers/callbacks)
    "ARG002", # unused method argument
    "ARG005", # unused lambda argument (common in callbacks)
    "SIM105", # use contextlib.suppress (explicit try-except is clearer)
    "SIM108", # use ternary operator (readability preference)
    "SIM102", # collapsible if (readability preference)
    "SIM116", # use dict instead of if-else (readability preference)
    "SIM117", # multiple with statements (readability preference)
]

[tool.ruff.lint.per-file-ignores]
"drift/core/drift_sdk.py" = ["F401"]  # unused imports for availability checking
"drift/instrumentation/**/e2e-tests/**" = ["F401", "F841"]  # test files

[tool.ruff.lint.isort]
known-first-party = ["drift"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ty.environment]
python-version = "3.9"

[tool.ty.src]
# Exclude e2e-tests directories from type checking
exclude = ["**/e2e-tests/**"]

[[tool.ty.overrides]]
# Disable unresolved-import errors for instrumentation files with optional dependencies
include = [
    "drift/instrumentation/django/**",
    "drift/instrumentation/grpc/**",
    "drift/instrumentation/psycopg/**",
    "drift/instrumentation/psycopg2/**",
    "drift/instrumentation/pyjwt/**",
    "drift/instrumentation/redis/**",
    "drift/instrumentation/kinde/**",
    "drift/instrumentation/http/transform_engine.py",
]

[tool.ty.overrides.rules]
"unresolved-import" = "ignore"
